feat: Initial minimum viable product
This includes the basic retrieval API, as well as view implementations for
viewing the data in the database. The data in the `seed` directory is an
initial set of data. The basic view also allows for import and delete of
the entire database. Views for editing values is to be added in a future
commit. Note that there is no security created yet, so this project is not
yet suitable for public hosting.